socket setup
