+++ build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     id("java")
+    kotlin("jvm")
 }
 
 group = "com.htilssu"
@@ -12,6 +13,7 @@ repositories {
 dependencies {
     testImplementation(platform("org.junit:junit-bom:5.10.0"))
     testImplementation("org.junit.jupiter:junit-jupiter")
+    implementation(kotlin("stdlib-jdk8"))
 }
 
 tasks.test {
@@ -19,4 +21,7 @@ tasks.test {
 }
 tasks.compileJava {
     options.encoding = "UTF-8"
+}
+kotlin {
+    jvmToolchain(21)
 }
