+++ src/main/java/com/htilssu/component/Position.java
@@ -1,0 +1,11 @@
+package com.htilssu.component;
+
+public class Position {
+    public int x;
+    public int y;
+
+    public Position(int x, int y) {
+        this.x = x;
+        this.y = y;
+    }
+}
