+++ src/main/java/com/htilssu/screen/AttackGrid.java
@@ -1,0 +1,5 @@
+package com.htilssu.screen;
+
+//Màn hình tấn công thuyền
+public class AttackGrid {
+}
